% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{aids_australia}
\alias{aids_australia}
\title{AIDS reported cases in Australia (1982-1988)}
\format{A time-series data structure.}
\source{
Daley D, Gani J (1999) Epidemic modelling: an introduction. Cambridge University
Press, Cambridge, MA.

Wilson S.R. (1989) Extrapolation forecasting. In Predicting the Course of AIDS
in Australia and Evaluating the Effect of AZT: A First Report, by Solomon P.J.,
Doust J.A. and Wilson S.R., National Centre for Epidemiology and Public Health,
Working Paper 3, Australian National University, Canberra, 33-41.
}
\usage{
aids_australia
}
\description{
A data set containing the number of reported AIDS case per month
in the UK from 1982 to 1990.
}
\examples{
data(aids_australia)
plot(aids_australia,type="o",pch=19,lty=2,xlab="time",ylab="incidence")
}
\keyword{datasets}

