% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{aids_uk}
\alias{aids_uk}
\title{AIDS reported cases in the UK (1982-1986)}
\format{A time-series data structure.}
\source{
Healy M., Tillett H. (1988) Short-term extrapolation of the AIDS epidemic.
J. Roy. Stat. Soc. A 151, 50-61.
}
\usage{
aids_uk
}
\description{
A data set containing the number of reported AIDS case per month
in the UK from 1982 to 1990.
}
\examples{
data(aids_uk)
plot(aids_uk,type="o",pch=19,lty=2,xlab="time",ylab="incidence")
}
\keyword{datasets}

