% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{aids_usa}
\alias{aids_usa}
\title{AIDS reported cases in the USA (1982-1990)}
\format{A time-series data structure.}
\source{
Hay, J.W. and Wolak, F.A. (1994) A procedure for estimating the unconditional
cumulative incidence curve and its variability for the human immunodeficiency
virus. Journal of the Royal Statistical Society C43, 599-624.

Rosenberg, P.S. (1990) A simple correction of AIDS surveillance data for
reporting delays. J. Acq. Immune Def. Synd. 3, 49-54.
}
\usage{
aids_usa
}
\description{
A data set containing the number of reported AIDS case per quarter
in the USA from 1982 to 1990.
}
\examples{
data(aids_usa)
plot(aids_usa,type="o",pch=19,lty=2,xlab="time",ylab="incidence")
}
\keyword{datasets}

