% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/midp.R
\docType{package}
\name{midp}
\alias{midp}
\alias{midp-package}
\title{midp: Modeling infectious diseases parameters based on serological and social
contact data.}
\description{
The midp package is based on the book "Modeling infectious diseases
parameters based on serological and social contact data" by Niel Hens, Ziv
Shkedy, Marc Aerts, Christel Faes, Pierre Van Damme and Philippe Beutels (
Springer, 2012).
}
\section{midp functions}{

The foo functions ...
}

